fixed meaningless usage of UnturnedPlayerComponent
