+++ UIElementsUnturned.UIElementsLibPluginExample/UIElementsUnturned.UIElementsLibPluginExample.csproj
@@ -79,7 +79,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Player\Components\PlayerUIListenerComponent.cs" />
     <Compile Include="Configuration\UIElementsLibPluginExampleConfiguration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UIElementsLibPluginExample.cs" />
